[BLINK-17983677][table] struct TableConfig
Summary: ref T17983677

Reviewers: 晓令

Reviewed By: 晓令

Subscribers: P710480

Differential Revision: https://aone.alibaba-inc.com/code/D803146

+++ flink-libraries/flink-table/src/test/java/org/apache/flink/table/resource/batch/calculator/DefaultParallelismCalculatorTest.java
@@ -52,7 +52,7 @@ public class DefaultParallelismCalculatorTest {
 		tableConfig = new TableConfig();
 		mq = mock(RelMetadataQuery.class);
 		tableConfig.getParameters().setString(TableConfig.SQL_EXEC_INFER_RESOURCE_MODE(), ExecResourceUtil.InferMode.ONLY_SOURCE.toString());
-		tableConfig.getParameters().setInteger(TableConfig.SQL_EXEC_INFER_RESOURCE_ROWS_PER_PARTITION(), 100);
+		tableConfig.getParameters().setLong(TableConfig.SQL_EXEC_INFER_RESOURCE_ROWS_PER_PARTITION(), 100);
 		tableConfig.getParameters().setInteger(TableConfig.SQL_EXEC_DEFAULT_PARALLELISM(), 50);
 		when(mq.getRowCount(tableSourceScan)).thenReturn(3000d);
 		when(mq.getAverageRowSize(tableSourceScan)).thenReturn(4d);
